refactor: deleted masterMembership package
